pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "redun-cert-exp"
version = "0.0.1"
description = "Experiments with certifiers for SDP relaxations with redundant constraints"
requires-python = ">=3.11,<3.13"
dependencies = [
	"numpy<2",
	"scipy",
	"scikit-sparse==0.4.16",
	"casadi",
	"cvxpy",
	"mosek",
	"matplotlib",
	"pandas",
	"pybind11",
	"pybind11-stubgen",
	"scikit-build-core",
	"cmake",
	"ninja",
	"pytest",
	"torch==2.5.1",
	"torchvision==0.20.1",
	"torchelastic",
	"asrl-pylgmath>=1.0.3",
	"debugpy>=1.8.20",
	"ranktools",
	"clipperpy",
	"cert-tools",
	"theseus-ai>=0.2.3",
	"gtsam-develop>=4.3a1.dev202604152020",
]

[project.optional-dependencies]
test = ["pytest"]

[tool.uv]
torch-backend = "auto"

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchaudio = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }
ranktools = { path = "../ranktools", editable = true }
clipperpy = { workspace = true }
cert-tools = { workspace = true, editable = true }

[tool.uv.workspace]
members = [
    "extern/clipper",
    "extern/certifiable-tools",
]

[tool.hatch.build.targets.wheel]
packages = ["src/max_clique","src/mat_weight_loc","src/utils"]
